config/shared: move validations into a subfolder
This mirrors the 'errors' move
